requirements.txt
python-dotenv
requests
selenium
webdriver_manager
geopy
fake-useragent
beautifulsoup4
supabase